feat(cubejs-cli): Move create/generate commands to TypeScript
